+++ packages/components/color-picker/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @yamada-ui/color-picker
 
+## 1.4.12
+
+### Patch Changes
+
+- [#3598](https://github.com/yamada-ui/yamada-ui/pull/3598) [`553772d`](https://github.com/yamada-ui/yamada-ui/commit/553772dcc1364dd2e2badb7935739ca5147ac0bd) Thanks [@kaehehehe](https://github.com/kaehehehe)! - remove `is` prefix from props
+
+- Updated dependencies [[`37b9d86`](https://github.com/yamada-ui/yamada-ui/commit/37b9d869638fb13983f53fd1aef1c45a8aa8e4df), [`4065286`](https://github.com/yamada-ui/yamada-ui/commit/4065286fe311de83fae4c203bf889de0789286f0), [`323e9b5`](https://github.com/yamada-ui/yamada-ui/commit/323e9b57b0925803068d9aff5802f2ae7cac1484), [`696da80`](https://github.com/yamada-ui/yamada-ui/commit/696da80ccd8adb30480fe34b6547ac2c81842f18), [`37b9d86`](https://github.com/yamada-ui/yamada-ui/commit/37b9d869638fb13983f53fd1aef1c45a8aa8e4df), [`bdb9cb3`](https://github.com/yamada-ui/yamada-ui/commit/bdb9cb3c4b0614405dced07ea333a4a698f5a189), [`3a74b00`](https://github.com/yamada-ui/yamada-ui/commit/3a74b003124a864d9882f541240cb9fccb9ef514), [`323e9b5`](https://github.com/yamada-ui/yamada-ui/commit/323e9b57b0925803068d9aff5802f2ae7cac1484), [`fcace07`](https://github.com/yamada-ui/yamada-ui/commit/fcace077aa6aac05e6c32d6592da956ab147d213), [`be42c26`](https://github.com/yamada-ui/yamada-ui/commit/be42c26cf3499335045d9a85c219dae164e656b2), [`c9e86f6`](https://github.com/yamada-ui/yamada-ui/commit/c9e86f6dd89bfef80a146d2b879321ec6923b824), [`604208c`](https://github.com/yamada-ui/yamada-ui/commit/604208c13d90f0c665616c7bd8c0f097b3c1b498), [`f86a6b2`](https://github.com/yamada-ui/yamada-ui/commit/f86a6b2fd3da6acde048e32e895f364bdb1d0f26), [`f86a6b2`](https://github.com/yamada-ui/yamada-ui/commit/f86a6b2fd3da6acde048e32e895f364bdb1d0f26)]:
+  - @yamada-ui/core@1.16.0
+  - @yamada-ui/portal@1.0.25
+  - @yamada-ui/button@1.0.52
+  - @yamada-ui/utils@1.6.0
+  - @yamada-ui/popover@1.4.5
+  - @yamada-ui/form-control@2.1.9
+  - @yamada-ui/icon@1.1.15
+  - @yamada-ui/input@1.0.49
+  - @yamada-ui/use-controllable-state@1.0.24
+  - @yamada-ui/use-disclosure@1.1.1
+  - @yamada-ui/use-eye-dropper@1.0.22
+  - @yamada-ui/use-outside-click@1.0.24
+  - @yamada-ui/use-pan-event@1.0.26
+  - @yamada-ui/use-size@1.0.24
+
 ## 1.4.11
 
 ### Patch Changes